+++ FRONT/src/components/App/index.js
@@ -10,8 +10,10 @@ import { Route, Routes } from 'react-router-dom';
 // == Import
 
 import Header from 'src/containers/Header';
+import GalerieHeader from 'src/components/Header/galerieHeader';
 import Home from 'src/containers/Home';
 import Biography from 'src/containers/Biography';
+import Galerie from 'src/containers/Galerie';
 import Footer from 'src/containers/Footer';
 
 import './app.scss';
@@ -26,10 +28,10 @@ const App = ({
 
   return (
     <div className="app">
-      <Header />
       <Routes>
-        <Route path="/" element={<Home />} />
-        <Route path="/biographie" element={<Biography />} />
+        <Route path="/" element={<><Header /><Home /></>} />
+        <Route path="/biographie" element={<><Header /><Biography /></>} />
+        <Route path="/galerie" element={<><GalerieHeader /><Galerie /></>} />
       </Routes>
       <Footer />
     </div>
