+++ FRONT/src/components/Galerie/galerie.scss
@@ -1,75 +1,4 @@
 @use 'src/styles/_vars';
 
 #galerie {
-  border: vars.$boxBorder;
-  border-radius: vars.$boxBorderRadius;
-  padding: 2rem;
-  display: none;
-  flex-direction: column;
-  justify-content: space-around;
-  
-  @media (min-width: vars.$inter) {
-    display: flex;
-  }
-
-  >p {
-    border: vars.$boxBorder;
-    border-radius: vars.$boxBorderRadius;
-    padding: 0.3rem 1.5rem;
-    margin: 0 auto;
-    width: fit-content;
-
-    &:not(:nth-child(3)) {
-      margin-bottom: 1rem;
-    }
-
-    &:hover {
-      color: vars.$textHoverColor;
-      text-decoration: vars.$textHoverDecoration;
-      box-shadow: vars.$boxHoverShadow;
-    }
-  }
-
-  >a {
-    margin: 0 auto;
-
-    &:hover {
-      color: vars.$textHoverColor;
-      text-decoration: vars.$textHoverDecoration;
-    }
-  }
-
-  ul {
-    text-align: center;
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-
-    li {
-      margin: 2rem 0;
-
-      a {
-        img {
-          max-height: 150px;
-          max-width: 150px;
-          margin: 0 auto;
-          box-shadow: vars.$imageShadow;
-          border: solid 3px white;
-          margin-bottom: 0.2rem;
-        }
-        
-        &:hover {
-          p {
-            color: vars.$textHoverColor;
-            text-decoration: vars.$textHoverDecoration;
-          }
-
-          img {
-            box-shadow: 3px 3px 3px vars.$secondaryColor;
-          }
-        }
-      }
-    }
-  }
 }
